file-classifier/requirements.txt
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5 
boto3==1.33.13
botocore==1.33.13
confluent-kafka==2.10.0
python-dotenv==1.2.1
requests==2.32.5
PyMuPDF==1.26.7
imagesize==1.4.1